feat: api 중간 작성

+++ backend_dev/src/main/java/com/ssafy/meongnyang/api/controller/CommentController.java
@@ -17,21 +17,25 @@ public class CommentController {
 
     @PostMapping
     public ResponseEntity<?> writeComment(@RequestBody CommentRegisterDto commentRegisterDto) {
-        return new ResponseEntity<>(new Response<>(true, 201, "댓글 등록 성공", commentService.writeComment(commentRegisterDto)), HttpStatus.CREATED);
+        return new ResponseEntity<>(new Response<>(true, 201, "댓글 등록 성공",
+                commentService.writeComment(commentRegisterDto)), HttpStatus.CREATED);
     }
 
     @PutMapping
     public ResponseEntity<?> updateComment(@RequestBody CommentUpdateDto commentUpdateDto) {
-        return new ResponseEntity<>(new Response<>(true, 202, "댓글 수정 성공", commentService.updateComment(commentUpdateDto)), HttpStatus.ACCEPTED);
+        return new ResponseEntity<>(new Response<>(true, 202, "댓글 수정 성공",
+                commentService.updateComment(commentUpdateDto)), HttpStatus.ACCEPTED);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteComment(@PathVariable int id) {
-        return new ResponseEntity<>(new Response<>(true, 200, "댓글 삭제 성공", commentService.deleteComment(id)), HttpStatus.OK);
+    public ResponseEntity<?> deleteComment(@PathVariable long id) {
+        return new ResponseEntity<>(new Response<>(true, 200, "댓글 삭제 성공",
+                commentService.deleteComment(id)), HttpStatus.OK);
     }
 
     @GetMapping("/list/{id}")
-    public ResponseEntity<?> getCommentList(@PathVariable int id) {
-        return new ResponseEntity<>(new Response<>(true, 200, "댓글 목록 조회 성공", commentService.getCommentList(id)), HttpStatus.OK);
+    public ResponseEntity<?> getCommentList(@PathVariable long id) {
+        return new ResponseEntity<>(new Response<>(true, 200, "댓글 목록 조회 성공",
+                commentService.getCommentList(id)), HttpStatus.OK);
     }
 }
